test(server): input transpilation morph relation
