+++ packages/twenty-server/src/engine/metadata-modules/flat-field-metadata/utils/from-morph-relation-create-field-input-to-flat-field-metadatas.util.ts
@@ -37,7 +37,10 @@ export const fromMorphRelationCreateFieldInputToFlatFieldMetadatas = async ({
   const rawMorphCreationPayload =
     createFieldInput.morphRelationsCreationPayload;
 
-  if (!isDefined(rawMorphCreationPayload)) {
+  if (
+    !isDefined(rawMorphCreationPayload) ||
+    rawMorphCreationPayload.length === 0
+  ) {
     return {
       status: 'fail',
       error: {
